requirements.txt
espn_api
jinja2
python-dotenv
click
black
requests
beautifulsoup4
pyobjson
tornado
